arm64: Fix the incompatible pointer type build error for user_cache_maint_handler()
Refer to commit 8d56e5c5a99c("arm64: Treat ESR_ELx as a 64-bit register"),
modify the parameter esr to unsigned long, otherwise it will
be detected by the compiler as an error.

+++ arch/arm64/kernel/traps.c
@@ -561,7 +561,7 @@ NOKPROBE_SYMBOL(do_ptrauth_fault);
 	} while (0)
 
 extern bool TKT340553_SW_WORKAROUND;
-static void user_cache_maint_handler(unsigned int esr, struct pt_regs *regs)
+static void user_cache_maint_handler(unsigned long esr, struct pt_regs *regs)
 {
 	unsigned long tagged_address, address;
 	int rt = ESR_ELx_SYS64_ISS_RT(esr);
